+++ Project1/Project1/IOClass.h
@@ -6,6 +6,6 @@ class IOClass
 public:
 	IOClass() {};
 	int** read(std::string name, int&n);
-	void write(std::string name, int**matrix, int n);
+	void write(std::string name, float**matrix, int n);
 };
